Finished the working webscraping code. Need to update NewLeafScraper with working Selenium code.
I completed this code a while ago (a few months ago), but never pushed it into GitHub. I was having a lot of issues with Selenium not scraping the Javascript "spoiler" content pop-up windows for a while. Now I need to just update the code for the other webpage (New Leaf) and I should have all the working code I need for text analysis.